requirements.txt
fastapi
uvicorn
pillow
sentence-transformers
torch

